+++ packages/devextreme/js/__internal/scheduler/appointments_new/__mock__/base_appointment_view.ts
